Implemented form creation for automatic tool installation.
This update includes the addition of a button that allows users to create forms for selecting tool type, tool, and version of software apps that a user would like to have. Backend functionality for the actual installation of tools will be added in future commits. Additionally, a validation check has been added to ensure all selections are properly made.